Enhance NotNullPtr

+++ src/IO/BTCapturer.cpp
@@ -31,7 +31,7 @@ IO::BTCapturer::BTCapturer(Bool autoStore) : thread(CheckThread, this, CSTR("BTC
 	this->lastFileName = 0;
 	this->autoStore = autoStore;
 	this->bt = IO::BTScanner::CreateScanner();
-	if (this->bt)
+	if (this->bt.NotNull())
 	{
 		//this->bt->SetScanMode(IO::BTScanner::SM_ACTIVE);
 	}
@@ -40,17 +40,18 @@ IO::BTCapturer::BTCapturer(Bool autoStore) : thread(CheckThread, this, CSTR("BTC
 IO::BTCapturer::~BTCapturer()
 {
 	this->Stop();
-	if (this->bt)
+	NotNullPtr<IO::BTScanner> bt;
+	if (this->bt.SetTo(bt))
 	{
-		this->bt->Close();
-		DEL_CLASS(this->bt);
+		bt->Close();
+		this->bt.Delete();
 	}
 	SDEL_TEXT(this->lastFileName);
 }
 
 Bool IO::BTCapturer::IsError()
 {
-	return this->bt == 0;
+	return this->bt.IsNull();
 }
 
 Bool IO::BTCapturer::IsStarted()
@@ -60,66 +61,74 @@ Bool IO::BTCapturer::IsStarted()
 
 Bool IO::BTCapturer::Start()
 {
-	if (this->IsError() || this->IsStarted())
+	NotNullPtr<IO::BTScanner> bt;
+	if (!this->bt.SetTo(bt) || this->IsStarted())
 	{
 		return false;
 	}
 	this->thread.Start();
-	this->bt->ScanOn();
+	bt->ScanOn();
 	return true;
 }
 
 void IO::BTCapturer::Stop()
 {
 	//printf("BTCapturer: Stopping\r\n");
-	if (this->thread.IsRunning())
+	NotNullPtr<IO::BTScanner> bt;
+	if (this->thread.IsRunning() && this->bt.SetTo(bt))
 	{
 		this->thread.Stop();
 		//printf("BTCapturer: Stopping 2\r\n");
-		this->bt->ScanOff();
+		bt->ScanOff();
 	}
 	//printf("BTCapturer: Stopping 3\r\n");
 }
 
 void IO::BTCapturer::StoreStatus()
 {
-	UTF8Char sbuff[512];
-	UTF8Char *sptr;
-	IO::BTDevLog btLog;
-	Sync::MutexUsage mutUsage;
-	btLog.AppendList(this->bt->GetPublicMap(mutUsage));
-	btLog.AppendList(this->bt->GetRandomMap(mutUsage));
-	UOSInt i;
-	Data::DateTime dt;
-	dt.SetCurrTime();
-	sptr = IO::Path::GetProcessFileName(sbuff);
-	i = Text::StrLastIndexOfCharC(sbuff, (UOSInt)(sptr - sbuff), IO::Path::PATH_SEPERATOR);
-	sptr = &sbuff[i + 1];
-	sptr = Text::StrConcatC(sptr, UTF8STRC("bt"));
-	sptr = dt.ToString(sptr, "yyyyMMddHHmmss");
-	sptr = Text::StrConcatC(sptr, UTF8STRC(".txt"));
-	if (btLog.StoreFile(CSTRP(sbuff, sptr)))
+	NotNullPtr<IO::BTScanner> bt;
+	if (this->bt.SetTo(bt))
 	{
-		if (this->lastFileName)
+		UTF8Char sbuff[512];
+		UTF8Char *sptr;
+		IO::BTDevLog btLog;
+		Sync::MutexUsage mutUsage;
+		btLog.AppendList(bt->GetPublicMap(mutUsage));
+		btLog.AppendList(bt->GetRandomMap(mutUsage));
+		UOSInt i;
+		Data::DateTime dt;
+		dt.SetCurrTime();
+		sptr = IO::Path::GetProcessFileName(sbuff);
+		i = Text::StrLastIndexOfCharC(sbuff, (UOSInt)(sptr - sbuff), IO::Path::PATH_SEPERATOR);
+		sptr = &sbuff[i + 1];
+		sptr = Text::StrConcatC(sptr, UTF8STRC("bt"));
+		sptr = dt.ToString(sptr, "yyyyMMddHHmmss");
+		sptr = Text::StrConcatC(sptr, UTF8STRC(".txt"));
+		if (btLog.StoreFile(CSTRP(sbuff, sptr)))
 		{
-			IO::Path::DeleteFile(this->lastFileName);
-			Text::StrDelNew(this->lastFileName);
+			if (this->lastFileName)
+			{
+				IO::Path::DeleteFile(this->lastFileName);
+				Text::StrDelNew(this->lastFileName);
+			}
+			this->lastFileName = Text::StrCopyNewC(sbuff, (UOSInt)(sptr - sbuff)).Ptr();
 		}
-		this->lastFileName = Text::StrCopyNewC(sbuff, (UOSInt)(sptr - sbuff)).Ptr();
 	}
 }
 
-NotNullPtr<const Data::ReadingList<IO::BTScanLog::ScanRecord3*>> IO::BTCapturer::GetPublicList(NotNullPtr<Sync::MutexUsage> mutUsage) const
+NotNullPtr<const Data::ReadingListNN<IO::BTScanLog::ScanRecord3>> IO::BTCapturer::GetPublicList(NotNullPtr<Sync::MutexUsage> mutUsage) const
 {
-	return this->bt->GetPublicMap(mutUsage);
+	return this->bt.OrNull()->GetPublicMap(mutUsage);
 }
 
-NotNullPtr<const Data::ReadingList<IO::BTScanLog::ScanRecord3*>> IO::BTCapturer::GetRandomList(NotNullPtr<Sync::MutexUsage> mutUsage) const
+NotNullPtr<const Data::ReadingListNN<IO::BTScanLog::ScanRecord3>> IO::BTCapturer::GetRandomList(NotNullPtr<Sync::MutexUsage> mutUsage) const
 {
-	return this->bt->GetRandomMap(mutUsage);
+	return this->bt.OrNull()->GetRandomMap(mutUsage);
 }
 
 void IO::BTCapturer::SetUpdateHandler(IO::BTScanner::RecordHandler hdlr, void *userObj)
 {
-	this->bt->HandleRecordUpdate(hdlr, userObj);
+	NotNullPtr<IO::BTScanner> bt;
+	if (this->bt.SetTo(bt))
+		bt->HandleRecordUpdate(hdlr, userObj);
 }
